+++ crates/resvg/examples/custom_href_resolver.rs
@@ -4,14 +4,18 @@ fn main() {
     let ferris_image = std::sync::Arc::new(std::fs::read("./examples/ferris.png").unwrap());
 
     // We know that our SVG won't have DataUrl hrefs, just return None for such case.
-    let resolve_data = Box::new(|_: &str, _: std::sync::Arc<Vec<u8>>, _: &usvg::Options| None);
+    let resolve_data = Box::new(
+        |_: &str, _: std::sync::Arc<Vec<u8>>, _: &usvg::Options, _: &usvg::fontdb::Database| None,
+    );
 
     // Here we handle xlink:href attribute as string,
     // let's use already loaded Ferris image to match that string.
-    let resolve_string = Box::new(move |href: &str, _: &usvg::Options| match href {
-        "ferris_image" => Some(usvg::ImageKind::PNG(ferris_image.clone())),
-        _ => None,
-    });
+    let resolve_string = Box::new(
+        move |href: &str, _: &usvg::Options, _: &usvg::fontdb::Database| match href {
+            "ferris_image" => Some(usvg::ImageKind::PNG(ferris_image.clone())),
+            _ => None,
+        },
+    );
 
     // Assign new ImageHrefResolver option using our closures.
     opt.image_href_resolver = usvg::ImageHrefResolver {